Additional Examples added

+++ LightParameter/LightParameter.h
@@ -60,8 +60,8 @@ unsigned long stopTime;
 unsigned long cycles;             //no of times the pattern should run
 
 int index;
-int onTime;
-int offTime;
+unsigned long onTime;
+unsigned long offTime;
 int totalsteps;        //no of iterations
 int brightness;          //the current brightness of the LED's
 
@@ -81,15 +81,17 @@ int grouplength;        //length of the group
 
 LightParameter();
 
-LightParameter(ActivePattern Pattern, Direction dir, int start, int cycle, int Index,
-                                int on, int off, int Brightness, uint32_t color1, uint32_t color2,
+LightParameter(ActivePattern Pattern, Direction dir, unsigned long start, int cycle, int Index,
+                                unsigned long on, unsigned long off, int Brightness, uint32_t color1, uint32_t color2,
                                 unsigned long Interval, int g[], int length);
 /*LightParameter(ActivePattern Pattern, Direction dir, int start, int cycle, int Index, int Brightness, unsigned long Interval, int g[], int length);
 LightParameter(ActivePattern Pattern, Direction dir, int start, int cycle, int Index, int Brightness, uint32_t color1, unsigned long Interval, int g[], int length);
 LightParameter(ActivePattern Pattern, Direction dir, int start, int cycle, int Index, int Brightness, uint32_t color1, uint32_t color2, unsigned long Interval, int g[], int length);
 LightParameter(ActivePattern Pattern, Direction dir, int start, int cycle, int on, int off, int Index, int Brightness, uint32_t color1, unsigned long Interval, int g[], int length);*/
 
-void initialize(ActivePattern Pattern, Direction dir, int start, int cycle, int Index, int on, int off, int Brightness, uint32_t color1, uint32_t color2, unsigned long Interval, int g[], int length);
+void initialize(ActivePattern Pattern, Direction dir, unsigned long start, int cycle, int Index,
+                unsigned long on, unsigned long off, int Brightness, uint32_t color1, uint32_t color2, unsigned long Interval,
+                int g[], int length);
 };
